move grid aggregator to submodule
