+++ modules/aggregation-layers/src/gpu-grid-layer/gpu-grid-layer.js
@@ -19,13 +19,12 @@
 // THE SOFTWARE.
 
 import {PhongMaterial} from '@luma.gl/core';
-import {
-  CompositeLayer,
-  _GPUGridAggregator as GPUGridAggregator,
-  _pointToDensityGridData as pointToDensityGridData
-} from '@deck.gl/core';
+import {CompositeLayer} from '@deck.gl/core';
 
 import GPUGridCellLayer from './gpu-grid-cell-layer';
+
+import GPUGridAggregator from '../utils/gpu-grid-aggregation/gpu-grid-aggregator';
+import {pointToDensityGridData} from '../utils/gpu-grid-aggregation/grid-aggregation-utils';
 
 const MINCOLOR = [0, 0, 0, 255];
 const MAXCOLOR = [0, 255, 0, 255];
