+++ modules/aggregation-layers/src/grid-layer/grid-layer.js
@@ -19,10 +19,12 @@
 // THE SOFTWARE.
 
 import {PhongMaterial} from '@luma.gl/core';
-import {CompositeLayer, experimental} from '@deck.gl/core';
-const {BinSorter, defaultColorRange, getQuantizeScale, getLinearScale} = experimental;
-
+import {CompositeLayer} from '@deck.gl/core';
 import {GridCellLayer} from '@deck.gl/layers';
+
+import BinSorter from '../utils/bin-sorter';
+import {defaultColorRange} from '../utils/color-utils';
+import {getQuantizeScale, getLinearScale} from '../utils/scale-utils';
 
 import {pointToDensityGridData} from './grid-aggregator';
