+++ modules/aggregation-layers/src/utils/bin-sorter.js
