+++ modules/aggregation-layers/src/utils/color-utils.js
