+++ modules/aggregation-layers/src/utils/gpu-grid-aggregation/gpu-grid-aggregator-constants.js
