+++ test/modules/aggregation-layers/index.js
@@ -24,3 +24,4 @@ import './hexagon-layer.spec';
 import './hexagon-aggregator.spec';
 import './contour-layer/marching-squares.spec';
 import './contour-layer/contour-layer.spec';
+import './utils/scale-utils.spec';
