+++ test/modules/core/utils/index.js
@@ -25,5 +25,4 @@ import './flatten.spec';
 import './positions.spec';
 import './memoize.spec';
 import './array-utils.spec';
-import './scale-utils.spec';
 // import './compare-objects.spec';
