+++ test/modules/index.js
@@ -31,9 +31,9 @@ import './json';
 if (typeof document !== 'undefined') {
   require('./react');
   require('./main/bundle');
-  require('./core/experimental/utils/gpu-grid-aggregator.spec');
+  require('./aggregation-layers/utils/gpu-grid-aggregator.spec');
   // TODO - This is failing in headless browser test. Might be related to
   // https://github.com/uber/luma.gl/issues/906
-  // require('./core/experimental/utils/grid-aggregation-utils.spec');
+  // require('./aggregation-layers/utils/grid-aggregation-utils.spec');
   require('./core/lib/pick-layers.spec');
 }
